Correctly indent if in ex 02 device code
As of now, the closing curly brace of an if statement
inside the __raygen__renderFrame() function in the
devicePrograms.cu file of the second example is
indented two spaces less than it should be, making it
look like the __raygen__renderFrame() function's
closing curly brace at first sight. This commit
indents it correctly such that its role is not
misunderstood.